refactor: migrate toaster-forge shared crates
Extract and consolidate previously inline types, utilities, and service logic into dedicated `aster_forge_*` crates:

- Replace `SystemConfigVisibility`, `SystemConfigSource`, `SystemConfigValueType` with `ConfigVisibility`, `ConfigSource`, `ConfigValueType` from `aster_forge_config` in frontend types and all Rust call sites
- Move `ExternalAuthProviderKind` and `ExternalAuthProtocol` from `crate::types` to `aster_forge_external_auth`
- Delegate audit lifecycle (`init`, `flush`, `shutdown`) to `aster_forge_audit`; remove-exports from local audit module
- Migrate `TaskExecutionContext`, `TaskLease`, `TaskStepInfo`, `TaskStepSpecDispatchStats`, `TaskRetryClass`, `TaskLaneConfig`, and related helpers to `aster_forge_tasks`; remove local duplicates
- Replace `crate::metrics::configure_route` with `aster_forge_actix_observability::configure_prometheus_route` and add `record_health_report` using the forge metrics backend
- Move `TemplateVariableGroup`/`TemplateVariableItem` and template registry access to `aster_forge_mail`; remove wrapper structs and forwarding functions from config schema
- Replace `health::ping_database` with `aster_forge_db::ping_database`; broaden error bound to `impl Display`
- Inline `access_token_hash` in WOPI session using `aster_forge_crypto::sha256_hex` directly; remove wrapper and its test
- Remove `host_is_lopback` wrapper in CLI node; call `aster_forge_utils::net::is_loopback_host` directly and drop redundant unit test
- Map `TaskCoreError` variants to typed API error codes instead of generic internal error
- Add `details` field to `AdminSystemHealthComponent` from `aster_forge_runtime::HealthComponentDetail`
- Change `RemoteNodeHealthTestStats` counter fields from `usize` to `u64`
- Add `Display` impl for `BackgroundTaskKind`
- Update `TaskExecutionContext::new` call sites to pass renewal timeout duration as explicit parameter
- Update generated OpenAPI schema and frontend API types accordingly

+++ src/services/remote/remote_node.rs
@@ -110,10 +110,10 @@ pub struct TestRemoteNodeInput {
 
 #[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
 pub struct RemoteNodeHealthTestStats {
-    pub checked: usize,
-    pub healthy: usize,
-    pub failed: usize,
-    pub skipped: usize,
+    pub checked: u64,
+    pub healthy: u64,
+    pub failed: u64,
+    pub skipped: u64,
 }
 
 struct ProbedRemoteNode {
